projects/gemini-scholar-pipeline/requirements.txt
playwright>=1.40.0
jinja2>=3.0.0
markdown>=3.5.0
tqdm>=4.66.0
google-genai
python-dotenv
